move mockpachd's panic handler into a package; use in real and test pachds
