+++ src/internal/middleware/logging/BUILD.bazel
@@ -12,6 +12,8 @@ go_library(
     deps = [
         "//src/internal/errors",
         "//src/internal/log",
+        "//src/internal/middleware/auth",
+        "//src/internal/middleware/recovery",
         "//src/internal/pctx",
         "@com_github_fatih_camelcase//:camelcase",
         "@com_github_prometheus_client_golang//prometheus",
