+++ src/internal/testpachd/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//src/internal/middleware/auth",
         "//src/internal/middleware/errors",
         "//src/internal/middleware/logging",
+        "//src/internal/middleware/recovery",
         "//src/internal/middleware/validation",
         "//src/internal/pctx",
         "//src/internal/require",
@@ -37,8 +38,6 @@ go_library(
         "//src/transaction",
         "//src/version/versionpb",
         "@org_golang_google_grpc//:grpc",
-        "@org_golang_google_grpc//codes",
-        "@org_golang_google_grpc//status",
         "@org_golang_google_protobuf//types/known/emptypb",
         "@org_golang_x_sync//errgroup",
     ],
